docker: fix v1638 MSSIM available memory
See for details:
  - https://github.com/kgoldman/ibmswtpm2/pull/4

TL;DR the available version doesn't have enough memory with RSA
bumped to 3072 to be PC Client Spec compatible and it causes
issues with the RM. Also, the build never respected CFLAGS, so
just make it append CFLAGS to it's CCFLAGS so we can at least
append to the flag set and control these values.

Drop openssl include location, it was never used.

Signed-off-by: William Roberts <william.c.roberts@intel.com>