build: modularize build using m4
Add a makefile that takes .docker.m4 suffix files and processes them
with m4 to produce the actual dockerfile used to build the docker image.
This can be used for things like include() which is not supported in
dockerfile so we don't have to repeat all the stanzas throughout more
files.

File count is going to to start to grow as we add new architectures.

Signed-off-by: William Roberts <william.c.roberts@intel.com>