ubuntu-16.04: reconfigure python
pip3 install of the packages was experiencing a build failure. The
recomended procedure was to update pip3. Since we really just need 1
type of python and that python to be version 3.7 or higher, drop all the
various python, python3 and pip installs in the Dockerfile and just use
python3.7 and it's pip.

Update python3 to point to python3.7 so scripts looking for python3
executable have it.

Signed-off-by: William Roberts <william.c.roberts@intel.com>