github actions: build on all push/pull_request
For forks, we want them to run in their own github actions for building,
even if they just push to a branch ie a push action. We also want to
build for incoming PR's so also do it for pull_request.

However, for publishing, it should only happen on push requests for the
tpm2-software org, as you need secrets. This way forks will have full
CI without the failing publish.

Rename the files and the names of the jobs to match the intended use.

Signed-off-by: William Roberts <william.c.roberts@intel.com>