opensuse*: setup python
We cannot remove python2, as python2 is the default python for OpenSuse
and the way that AM_PYTHON_PATh works is configure.ac for tpm2-tools is
to find the first python version that is at least 2.7.

Add back python2, we only need YAML. Also, clean up how python is
installed by using the pip3 macro.

Signed-off-by: William Roberts <william.c.roberts@intel.com>