modules: merge RUN commands together
When building some images on a system with 60 GB free disk space, the
build fails with "no space left on device". This is caused by the large
number of layers which are created for each image.

Minimize the number of intermediate layers by combining all RUN
statements together in each module, when this is appropriate.

Also use `cd ...` instead of `WORKDIR` statements to also reduce the
number of intermediate layers.

With this, `podman build -f ./fedora-32.docker` succeeds on my test
system which has "only" 60 GB free disk space.

Signed-off-by: Nicolas Iooss <nicolas.iooss@ledger.fr>