README: add title, links and references to Podman, and update commands
First, add a title to README.md that states that this is the README of
`tpm2-software-container` project. Otherwise the README which is display
on pages such as
https://github.com/orgs/tpm2-software/packages/container/package/fedora-32
looks misplaced (the first sentence starts with "This repository" even
though it is associated with a Docker image...).

Secondly, add links to various places and document using Podman, which
can be used on systems where the user is not root (and thus should not
be in the `docker` group, as this is the same as being root according to
https://docs.docker.com/engine/install/linux-postinstall/#manage-docker-as-a-non-root-user ).

Finally, update the documented commands because `docker build .` does
not work: there is no `Dockerfile` in the git repository. The recipes
first needs to be generated using `make`, and then option `-f` has to be
used in order to specify which file to build.

Signed-off-by: Nicolas Iooss <nicolas.iooss@ledger.fr>