images: add images with openssl3.
The dev packages for curl and openssl are removed and
openssl3 and curl compiled with openssl3 are installed.
Because this will only temporary used until ossl3 is available on
distributions this workaround was chosen, because no changes on
the other m4 files were needed.
The imagess ubuntu-20.0, fedora-32, and opensuse-leap were
extended.

Signed-off-by: Michael Eckel <michael.eckel@sit.fraunhofer.de>
Signed-off-by: Juergen Repp <juergen.repp@sit.fraunhofer.de>