modules: ensure that pip sees all dependencies
When pinning a version with pip, if pip doesnt see all the installs at
once it tries to upgrade package versions even if they satisfy the
requirement. This causes issues since cryptograpy > 3.4.8 requires RUST
and some images don't have a new enough rust toolchain and thus the
wheel wont build. To fix this, run pip install and in one command with
all packages.

Signed-off-by: William Roberts <william.c.roberts@intel.com>