ci: ensure base ref for push actions
It seems as though the git diff output for the push event is only
diffin'g the head commit, which means that if a push contains N>1
patches, decisions on what to build are missing the full context of
changed files. This causes many containers to be missed in publishing to
GHCR.

To fix this, per issue [154], since version 4.2.0 of git-diff-action, it
supports the with BASE context to set the BASE ref on push. Set this to
use refs/heads/master so diffs have the full context.

Also support a manual build and push option so we can always trigger a
build and publish if CI bugs creep back in.

154. https://github.com/technote-space/get-diff-action/issues/154

Signed-off-by: William Roberts <william.c.roberts@intel.com>