ubuntu-22.04: replace swtpm with tpm_server
In multi-stage encrypt/decrypt, like pkcs11-crypt tests, the CI was
failing. I was able to debug this and swtpm and the current time is
failing as the last block coming back of plaintext is wrong. My guess is
some bug between it and this version of openssl 3.0. To save time, the
simple fix is to just use tpm_server, so do that, but don't add it do
modules as we don't want to re-trigger building of all images.

Signed-off-by: William Roberts <william.c.roberts@intel.com>